third_party/WebKit/LayoutTests/http/tests/serviceworker/cache-match-expected.txt

This is a testharness.js-based test.
PASS Service Worker: Cache.match and Cache.matchAll 
FAIL Cache.matchAll with no matching entries promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with no matching entries 
FAIL Cache.matchAll with URL promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with URL 
FAIL Cache.matchAll with Request promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with Request 
FAIL Cache.matchAll with new Request promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with new Request 
PASS Cache.match with Request containing non-empty body 
FAIL Cache.matchAll with ignoreSearch option (request with no search parameters) promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with ignoreSearch option (request with no search parameters) 
FAIL Cache.matchAll with ignoreSearch option (request with search parameter) promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with ignoreSearch option (request with search parameter) 
FAIL Cache.matchAll with URL containing fragment promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with URL containing fragment 
FAIL Cache.matchAll with string fragment "http" as query promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with string fragment "http" as query 
FAIL Cache.matchAll with prefixMatch option promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with prefixMatch option 
FAIL Cache.matchAll with prefixMatch option (URL ending with path delimiter) promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
FAIL Cache.match with prefixMatch option (URL ending with path delimiter) assert_equals: Cache.match should honor prefixMatch.: Response expected (object) object "[object Response]" but got (undefined) undefined
FAIL Cache.matchAll with prefixMatch option (URL with embedded credentials) promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
FAIL Cache.match with prefixMatch option (URL with embedded credentials) assert_equals: Cache.match should honor prefixMatch.: Response expected (object) object "[object Response]" but got (undefined) undefined
FAIL Cache.matchAll with prefixMatch option (URL matching embedded credentials) promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with prefixMatch option (URL matching embedded credentials) 
FAIL Cache.matchAll with URL containing credentials promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with URL containing credentials 
FAIL Cache.matchAll with responses containing "Vary" header promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
FAIL Cache.match with responses containing "Vary" header assert_true: Cache.match should honor "Vary" header. expected true got false
FAIL Cache.matchAll with "ignoreVary" parameter promise_test: Unhandled rejection with value: object "TypeError: undefined is not a function"
PASS Cache.match with Request and Response objects with different URLs 
PASS Cache.match invoked multiple times for the same Request/Response 
Harness: the test ran to completion.
